Refactoring code in Transaction and Account.
1)
Account.java depends on Transaction.java
Many of the signing operations are in Account.java
Given the dependency direction, Transaction cannot utilize the signing operations in Account.
Moving the signing operations to Transaction, so that both Transaction and Account can use them.
Since some of the signing operations need the account, factored out the core signing operations to utils.SignatureUtils.

2)
Added SignatureUtils.getDummyPrivateKey utility to generate a private key and cache it.
This is used for signing transactions for the purpose of measuring the encoding size to determine the fee.
Note that a new key is not generated every time when getDummyPrivateKey is called.

3)
Added:
acceptAssetTransaction
assetTransferTransaction

4)
Reformatted indentation in Transaction.java
Added missing comments to Transaction.java

5)
Modified ApiException to include resoponseBody text with the message.

6)
Deprecated transactionWithSuggestedFeePerByte
replaced by Transaction.setFeeWithSuggestedFeePerByte

7)
Adding tests to asset transfer and asset accept transactions.