1) Added asset xfer and accept transaction methods: createAcceptAssetTransaction createTransferAssetTransaction
2)
Modified some of the Transaction constructors not to call the other constructor used for creating Transactions from json.
Otherwise, every time a new field is added, all these constructors will need updating to pass null values.

5)
Modified ApiException to include resoponseBody text with the message. This is to enhance the error reporting downstream.

6)
Changed Account.transactionWithSuggestedFeePerByte not to return a new transaction, but to set the transaction fee on the passed transaction object. This corrects the shallow copy problem: it was returning a new transaction with shallow copied members from the original transaction.

7)
Added tests to asset transfer and asset accept transactions with signed golden.