Generator enhancements and go-sdk generator  (#157)
Generator enhancements and go-sdk generator

In this change, the generator is enhanced to be a general purpose
generator.

Publisher/Subscriber paradigm is added to build generators outside the
spec file parser.

Relevant events such as new model or new parameter, with the relevant
customizations, e.g. type conversion are published, and subscribers
can pick up the messages and generate code.

The first use case of the publisher/subscriber is implemented for
generating go-sdk code.


Valuable comments here. Will be revisited in future.